move helpers directory into the lib/db directory
