+++ __tests__/db/helpers/get-store-number.test.js
@@ -1,4 +1,4 @@
-const getStoreNumber = require('../../lib/helpers/get-store-number')
+const getStoreNumber = require('../../../lib/db/helpers/get-store-number')
 
 test('getStoreNumber: Should return the number from collection path', () => {
     let file = './users/users.1.json'
