+++ lib/db/Collection.js
@@ -1,5 +1,4 @@
 const getMetaFile = require('./metas/get-meta-file')
-const getCollectionModel = require('../helpers/get-model')
 const getDocById = require('./collection/get-document-by-id')
 const insertOneDocument = require('./collection/insert-one-document')
 const insertManyDocuments = require('./collection/insert-many-documents')
@@ -14,9 +13,10 @@ const setProperty = require('./collection/set-property')
 const insertIntoArray = require('./collection/insert-into-array')
 const removeFromArray = require('./collection/remove-from-array')
 const updatePropertyArray = require('./collection/update-property-array')
-const validateSetProperty = require('../helpers/validate-set-property')
-const validateDeleteProperty = require('../helpers/validate-delete-property')
-const validateInsertToValues = require('../helpers/validate-insert-to-values')
+const getCollectionModel = require('./helpers/get-model')
+const validateSetProperty = require('./helpers/validate-set-property')
+const validateDeleteProperty = require('./helpers/validate-delete-property')
+const validateInsertToValues = require('./helpers/validate-insert-to-values')
 
 module.exports = class Collection {
