+++ lib/db/collection/delete-many-documents.js
@@ -2,7 +2,7 @@ const fs = require('fs')
 const path = require('path')
 const { Readable, Transform, pipeline } = require('stream')
 const validate = require('../validate')
-const getStoreDeleteData = require('../../helpers/get-store-delete-data')
+const getStoreDeleteData = require('../helpers/get-store-delete-data')
 const updateManyStores = require('../metas/update-many-stores')
 
 class DeleteManyReadable extends Readable {
