+++ lib/db/collection/geo-search.js
@@ -1,4 +1,4 @@
-const setQueryParams = require('../../helpers/set-query-params')
+const setQueryParams = require('../helpers/set-query-params')
 
 const geoSearch = (data, params, geoParams) => {
     if (!Object.prototype.toString.call(geoParams) == '[object Object]' || !geoParams.lat || !geoParams.long || !geoParams.radius) {
