+++ lib/db/collection/get-all-documents.js
@@ -1,6 +1,6 @@
 const path = require('path')
 const { Readable } = require('stream')
-const getCollectionResources = require('../../helpers/get-col-resources')
+const getCollectionResources = require('../helpers/get-col-resources')
 
 class GetAllReadable extends Readable {
