+++ lib/db/collection/get-document-by-id.js
@@ -1,7 +1,7 @@
 const fs = require('fs')
 const { promisify } = require('util')
 const validate = require('../validate')
-const getDocStorePath = require('../../helpers/get-store-path')
+const getDocStorePath = require('../helpers/get-store-path')
 
 const readFile = promisify(fs.readFile)
