+++ lib/db/collection/insert-many-documents.js
@@ -1,7 +1,7 @@
-const isMaxSize = require('../../helpers/is-max-size')
+const isMaxSize = require('../helpers/is-max-size')
 const newStoreGen = require('../../generators/new-store-generator')
-const readFromStore = require('../../helpers/read-from-store')
-const writeToStore = require('../../helpers/write-to-store')
+const readFromStore = require('../helpers/read-from-store')
+const writeToStore = require('../helpers/write-to-store')
 const updateOneStore = require('../metas/update-one-store')
 const createManyDocs = require('../models/default/create-many-documents')
 const SchemaDocument = require('../models/schema/SchemaDocument')
