+++ lib/db/collection/insert-one-document.js
@@ -1,7 +1,7 @@
-const isMaxSize = require('../../helpers/is-max-size')
+const isMaxSize = require('../helpers/is-max-size')
 const splitStoreFile = require('../../generators/split-store-file')
-const readFromStore = require('../../helpers/read-from-store')
-const writeToStore = require('../../helpers/write-to-store')
+const readFromStore = require('../helpers/read-from-store')
+const writeToStore = require('../helpers/write-to-store')
 const updateOneStore = require('../metas/update-one-store')
 const createOneDoc = require('../models/default/create-one-document')
 const SchemaDocument = require('../models/schema/SchemaDocument')
