+++ lib/db/collection/query-collection.js
@@ -2,8 +2,8 @@ const path = require('path')
 const { Readable } = require('stream')
 const dotProp = require('dot-prop')
 const validate = require('../validate')
-const getCollectionResources = require('../../helpers/get-col-resources')
-const setQueryParams = require('../../helpers/set-query-params')
+const getCollectionResources = require('../helpers/get-col-resources')
+const setQueryParams = require('../helpers/set-query-params')
 const geoSearch = require('./geo-search')
 
 class QueryReadable extends Readable {
