+++ lib/db/collection/update-many-documents.js
@@ -2,7 +2,7 @@ const fs = require('fs')
 const path = require('path')
 const { Readable, pipeline } = require('stream')
 const validate = require('../validate')
-const getStoreUpdatesData = require('../../helpers/get-store-updates-data')
+const getStoreUpdatesData = require('../helpers/get-store-updates-data')
 const updateDoc = require('../models/default/update-document')
 const updateOneStore = require('../metas/update-one-store')
 const SchemaDocument = require('../models/schema/SchemaDocument')
