+++ lib/db/collection/update-one-document.js
@@ -3,7 +3,7 @@ const path = require('path')
 const { Readable } = require('stream')
 const validate = require('../validate')
 const getDocById = require('./get-document-by-id')
-const getDocStorePath = require('../../helpers/get-store-path')
+const getDocStorePath = require('../helpers/get-store-path')
 const updateDoc = require('../models/default/update-document')
 const updateOneStore = require('../metas/update-one-store')
 const SchemaDocument = require('../models/schema/SchemaDocument')
