+++ lib/db/collection/update-property-array.js
@@ -3,7 +3,7 @@ const path = require('path')
 const { Readable } = require('stream')
 const dotProp = require('dot-prop')
 const validate = require('../validate')
-const getCollectionResources = require('../../helpers/get-col-resources')
+const getCollectionResources = require('../helpers/get-col-resources')
 const updateOneStore = require('../metas/update-one-store')
 
 class UpdateArrayReadable extends Readable {
