+++ lib/db/helpers/deep-merge.js
