+++ lib/db/helpers/generate-id.js
@@ -1,5 +1,5 @@
 const { uid } = require('uid')
-const validate = require('../db/validate')
+const validate = require('../../db/validate')
 
 const generateId = {
     uid: (len) => {
