+++ lib/db/helpers/get-col-resources.js
