+++ lib/db/helpers/get-model-by-colname.js
