+++ lib/db/helpers/get-store-delete-data.js
@@ -1,4 +1,4 @@
-const validate = require('../db/validate')
+const validate = require('../../db/validate')
 
 // Get all store paths containing id's of docs to delete
 // returns [{ 'storePath', [idsArray] }]
