+++ lib/db/helpers/get-store-number.js
