+++ lib/db/helpers/get-store-path.js
@@ -1,4 +1,4 @@
-const validate = require('../db/validate')
+const validate = require('../../db/validate')
 
 // Get store path containing id
 const getDocStorePath = (colMeta, id) => {
