+++ lib/db/helpers/get-store-updates-data.js
@@ -1,4 +1,4 @@
-const validate = require('../db/validate')
+const validate = require('../../db/validate')
 
 // Get all store paths containing id's of docs to update
 // returns [{ 'storePath', [updates] }]
