+++ lib/db/helpers/is-max-size.js
