+++ lib/db/helpers/validate-delete-property.js
