+++ lib/db/helpers/validate-insert-to-values.js
