+++ lib/db/helpers/validate-set-property.js
@@ -1,4 +1,4 @@
-const updateOneDocument = require('../db/collection/update-one-document')
+const updateOneDocument = require('../../db/collection/update-one-document')
 
 const validateSetProperty = async (propertyPath, value, model) => {
     if (!model) {
