+++ lib/db/helpers/write-to-store.js
