+++ lib/db/metas/update-many-stores.js
@@ -2,7 +2,7 @@
 const fs = require('fs')
 const { promisify } = require('util')
 const writeFile = promisify(fs.writeFile)
-const getStoreNumber = require('../../helpers/get-store-number')
+const getStoreNumber = require('../helpers/get-store-number')
 
 /**
  * Update one or more collection meta store documents array of id's
