+++ lib/db/metas/update-one-store.js
@@ -1,7 +1,7 @@
 const fs = require('fs')
 const { promisify } = require('util')
 const writeFile = promisify(fs.writeFile)
-const getStoreNumber = require('../../helpers/get-store-number')
+const getStoreNumber = require('../helpers/get-store-number')
 
 /**
  * Update the collection meta store documents array of id's
