+++ lib/db/models/default/create-many-documents.js
@@ -1,5 +1,5 @@
 const validate = require('../../validate')
-const generateId = require('../../../helpers/generate-id')
+const generateId = require('../../helpers/generate-id')
 
 const createManyDocs = (colMeta, docObjects) => {
     validate.isObject(colMeta)
