+++ lib/db/models/default/create-one-document.js
@@ -1,6 +1,6 @@
 const validate = require('../../validate')
 const getDocById = require('../../collection/get-document-by-id')
-const generateId = require('../../../helpers/generate-id')
+const generateId = require('../../helpers/generate-id')
 
 const createOneDoc = async (colMeta, docObj) => {
     validate.isObject(colMeta)
