+++ lib/db/models/default/update-document.js
@@ -1,4 +1,4 @@
-const deepMerge = require('../../../helpers/deep-merge')
+const deepMerge = require('../../helpers/deep-merge')
 
 const updateDocument = (document, update) => {
     // compare doc updates, make changes
