+++ lib/db/models/schema/schema-types/SchemaString.js
@@ -86,7 +86,7 @@ class SchemaString extends SchemaType {
 
             if (param === 'validate') {
                 if (typeof ruleValue !== 'function') {
-                    throw new Error(`[schemaError]: 'validate' option must be an function, received: ${typeof ruleValue}`)
+                    throw new Error(`[schemaError]: 'validate' option must be a function, received: ${typeof ruleValue}`)
                 }
             }
         }
