+++ lib/generators/new-store-generator.js
@@ -1,10 +1,10 @@
 const fs = require('fs')
 const { promisify } = require('util')
 const writeFile = promisify(fs.writeFile)
-const getStoreNumber = require('../helpers/get-store-number')
-const isMaxSize = require('../helpers/is-max-size')
-const readFromStore = require('../helpers/read-from-store')
-const writeToStore = require('../helpers/write-to-store')
+const getStoreNumber = require('../db/helpers/get-store-number')
+const isMaxSize = require('../db/helpers/is-max-size')
+const readFromStore = require('../db/helpers/read-from-store')
+const writeToStore = require('../db/helpers/write-to-store')
 const updateOneStore = require('../db/metas/update-one-store')
 
 const newStoreGen = async (colMeta, documents) => {
