+++ lib/generators/split-store-file.js
@@ -1,7 +1,7 @@
 const fs = require('fs')
 const { promisify } = require('util')
 const writeFile = promisify(fs.writeFile)
-const getStoreNumber = require('../helpers/get-store-number')
+const getStoreNumber = require('../db/helpers/get-store-number')
 
 const splitStoreFile = async (colMeta, documents) => {
     try {
